move methods to liana.py
