+++ openproblems/tasks/cell_cell_communication/methods/__init__.py
@@ -1,6 +1,6 @@
-from .cellphonedb import cellphonedb
-from .connectome import connectome
+from .liana import cellphonedb
+from .liana import connectome
 from .liana import liana
-from .logfc import logfc
-from .natmi import natmi
-from .sca import sca
+from .liana import logfc
+from .liana import natmi
+from .liana import sca
